Split Api Routes into separate files
Remove skeleton submit_task threadpool
