+++ src/routes.rs
@@ -1,1 +1,5 @@
 pub mod info;
+pub mod peer;
+pub mod identity;
+pub mod wallet;
+pub mod transaction;
